+++ methods/src/test/kotlin/methods/equations/GmLinearEquationsTest.kt
@@ -1,4 +1,5 @@
 package methods.equations
+
 import engine.context.Preset
 import engine.methods.testMethodInX
 import org.junit.jupiter.api.Tag
@@ -8,7 +9,7 @@ import org.junit.jupiter.api.Test
 // The tag can be on the class as here, in which case all tests in the class will be tagged, or
 // on individual tests.
 @Tag("GmAction")
-class GmLinearEquationsTest {
+class GmOneStepLinearEquationMultiplicationTests {
     @Test
     fun `ax=c`() = testMethodInX(Preset.GMFriendly) {
         method = EquationsPlans.SolveEquation
@@ -44,6 +45,39 @@ class GmLinearEquationsTest {
         check { toExpr = "SetSolution[x : {1}]" }
     }
 
+    // to-do: Need to have [10/10]x ==> x
+    // @Test
+    // fun `frac{a}{b}x=c`() = testMethodInX(Preset.GMFriendly) {
+    //     method = EquationsPlans.SolveEquation
+    //     inputExpr = "[3/2]x = 5"
+    //     check { toExpr = "SetSolution[x : {[10/3]}]" }
+    // }
+
+    // to-do: Need to have [6x/3] ==> 2x and [10/10]x ==> x
+    // @Test
+    // fun `frac{a}{b}x=Na`() = testMethodInX(Preset.GMFriendly) {
+    //     method = EquationsPlans.SolveEquation
+    //     inputExpr = "[3/2]x = 6"
+    //     check { toExpr = "SetSolution[x : {4}]" }
+    // }
+    //
+    // @Test
+    // fun `frac{a}{b}x=frac{c}{d}`() = testMethodInX(Preset.GMFriendly) {
+    //     method = EquationsPlans.SolveEquation
+    //     inputExpr = "[3/2]x = [5/7]"
+    //     check { toExpr = "SetSolution[x : {[10/21]}]" }
+    // }
+    //
+    // @Test
+    // fun `frac{a}{b}x=frac{Ma}{Nb}`() = testMethodInX(Preset.GMFriendly) {
+    //     method = EquationsPlans.SolveEquation
+    //     inputExpr = "[3/2]x = [9/10]"
+    //     check { toExpr = "SetSolution[x : {[3/5]}]" }
+    // }
+}
+
+@Tag("GmAction")
+class GmOneStepLinearEquationDivisionTests {
     @Test
     fun `frac{x}{a}=b`() = testMethodInX(Preset.GMFriendly) {
         method = EquationsPlans.SolveEquation
@@ -119,6 +153,16 @@ class GmLinearEquationsTest {
     }
 
     @Test
+    fun `frac{x}{-a}=-b`() = testMethodInX(Preset.GMFriendly) {
+        method = EquationsPlans.SolveEquation
+        inputExpr = "[x/-3] = -4"
+        check { toExpr = "SetSolution[x : {12}]" }
+    }
+}
+
+@Tag("GmAction")
+class GmOneStepLinearEquationAdditionTests {
+    @Test
     fun `x+a=b`() = testMethodInX(Preset.GMFriendly) {
         method = EquationsPlans.SolveEquation
         inputExpr = "x+1 = 4"
@@ -166,25 +210,4 @@ class GmLinearEquationsTest {
         inputExpr = "x+3 = 3"
         check { toExpr = "SetSolution[x : {0}]" }
     }
-
-    @Test
-    fun `frac{x}{-a}=-b`() = testMethodInX(Preset.GMFriendly) {
-        method = EquationsPlans.SolveEquation
-        inputExpr = "[x/-3] = -4"
-        check { toExpr = "SetSolution[x : {12}]" }
-    }
-    // TO-DO: Adjust Solver so that [6x/3] ==> 2x, not [2*3x/3*1]
-    // @Test
-    // fun `frac{a}{b}x=Na`() = testMethodInX(Preset.GMFriendly) {
-    //     method = EquationsPlans.SolveEquation
-    //     inputExpr = "[3/5]x = 6"
-    //     check { toExpr = "SetSolution[x : {10}]" }
-    // }
-    // TO-DO: Adjust Solver so that 15/15*x ==> x, not 1*x
-    // @Test
-    // fun `frac{a}{b}x=c`() = testMethodInX(Preset.GMFriendly) {
-    //     method = EquationsPlans.SolveEquation
-    //     inputExpr = "[3/5]x = 7"
-    //     check { toExpr = "SetSolution[x : { [35/3] }]" }
-    // }
 }